// src/data/models/User.ts
// Définition d'une interface TypeScript pour représenter un utilisateur
export interface User {
    id: number;      // Identifiant unique de l'utilisateur
    name: string;    // Nom de l'utilisateur
    prenom: string;    // Nom de l'utilisateur
    email: string;   // Email de l'utilisateur
  }
  